add default sourceroot
webpack and this loader is weird.

if no source root is set then it assumes the root of the project and not the package so you would get some sources like `../src/hub.ts` instead of the actual path: `./node_modules/@senty/hub/src/hub.ts`

Setting the sourceRoot to default to the package root fixes this issue.